Unused code files un-included. Player.dm deleted.
The ops view files are un-included because they are related to interface
panes that have been removed. They have not been entirely deleted
because some of the functionality in the op interface may not yet be
accessible with simple op verbs. Player.dm was related to the games
interface which has also been removed, and the roll() commands are no
longer useable anyway.